api/requirements.txt
python-dotenv==1.0.0
quart==0.18.4
requests==2.30.0